Move Anthropic semantic-convention instrumentation into LoongSuite (#176)
