+++ instrumentation-loongsuite/README.md
@@ -3,6 +3,7 @@
 | --------------- | ------------------ | --------------- | -------------- |
 | [loongsuite-instrumentation-agentscope](./loongsuite-instrumentation-agentscope) | agentscope >= 1.0.0 | No | development
 | [loongsuite-instrumentation-agno](./loongsuite-instrumentation-agno) | agno | No | development
+| [loongsuite-instrumentation-anthropic](./loongsuite-instrumentation-anthropic) | anthropic >= 0.16.0 | No | development
 | [loongsuite-instrumentation-claude-agent-sdk](./loongsuite-instrumentation-claude-agent-sdk) | claude-agent-sdk >= 0.1.0 | No | development
 | [loongsuite-instrumentation-crewai](./loongsuite-instrumentation-crewai) | crewai >= 0.80.0 | No | development
 | [loongsuite-instrumentation-dashscope](./loongsuite-instrumentation-dashscope) | dashscope >= 1.0.0 | No | development
