+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/examples/manual/README.rst
@@ -1,0 +1,50 @@
+OpenTelemetry Anthropic Instrumentation Example
+===============================================
+
+This is an example of how to instrument Anthropic calls when configuring OpenTelemetry SDK and Instrumentations manually.
+
+When `main.py <main.py>`_ is run, it exports traces and logs to an OTLP
+compatible endpoint. Traces include details such as the model used and the
+duration of the chat request. Logs capture the chat request and the generated
+response, providing a comprehensive view of the performance and behavior of
+your Anthropic requests.
+
+Note: `.env <.env>`_ file configures additional environment variables:
+
+- `OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT=true` configures
+  Anthropic instrumentation to capture prompt and completion contents on
+  events.
+
+Setup
+-----
+
+An OTLP compatible endpoint should be listening for traces and logs on
+http://localhost:4317. If not, update "OTEL_EXPORTER_OTLP_ENDPOINT" as well.
+
+Next, set up a virtual environment like this:
+
+::
+
+    python3 -m venv .venv
+    source .venv/bin/activate
+    pip install "python-dotenv[cli]"
+    pip install -r requirements.txt
+
+You will also need an Anthropic API key. Set it as an environment variable:
+
+::
+
+    export ANTHROPIC_API_KEY=your_api_key_here
+
+Run
+---
+
+Run the example like this:
+
+::
+
+    dotenv run -- python main.py
+
+You should see a poem generated by Claude while traces and logs export to your
+configured observability tool.
+
