+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/examples/manual/main.py
@@ -1,0 +1,47 @@
+# pylint: skip-file
+import anthropic
+
+# NOTE: OpenTelemetry Python Logs and Events APIs are in beta
+from opentelemetry import _logs, trace
+from opentelemetry.exporter.otlp.proto.grpc._log_exporter import (
+    OTLPLogExporter,
+)
+from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import (
+    OTLPSpanExporter,
+)
+from opentelemetry.instrumentation.anthropic import AnthropicInstrumentor
+from opentelemetry.sdk._logs import LoggerProvider
+from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
+
+# configure tracing
+trace.set_tracer_provider(TracerProvider())
+trace.get_tracer_provider().add_span_processor(
+    BatchSpanProcessor(OTLPSpanExporter())
+)
+
+# configure logging and events
+_logs.set_logger_provider(LoggerProvider())
+_logs.get_logger_provider().add_log_record_processor(
+    BatchLogRecordProcessor(OTLPLogExporter())
+)
+
+# instrument Anthropic
+AnthropicInstrumentor().instrument()
+
+
+def main():
+    client = anthropic.Anthropic()
+    message = client.messages.create(
+        model="claude-3-5-sonnet-20241022",
+        max_tokens=1024,
+        messages=[
+            {"role": "user", "content": "Write a short poem on OpenTelemetry."}
+        ],
+    )
+    print(message.content[0].text)
+
+
+if __name__ == "__main__":
+    main()
