+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/examples/manual/requirements.txt
@@ -1,0 +1,4 @@
+anthropic>=0.16.0
+opentelemetry-sdk~=1.36.0
+loongsuite-instrumentation-anthropic # TODO: update to the released version when available
+opentelemetry-exporter-otlp-proto-grpc~=1.36.0
