+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/examples/zero-code/README.rst
@@ -1,0 +1,56 @@
+OpenTelemetry Anthropic Zero-Code Instrumentation Example
+=========================================================
+
+This is an example of how to use OpenTelemetry's automatic instrumentation
+(zero-code) capabilities with the Anthropic SDK.
+
+The `opentelemetry-instrument` CLI automatically instruments your Python
+application without requiring code changes. When `main.py <main.py>`_ is run
+with the CLI, it exports traces and logs to an OTLP compatible endpoint.
+
+Note: `.env <.env>`_ file configures additional environment variables:
+
+- `OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT=true` configures
+  Anthropic instrumentation to capture prompt and completion contents on
+  events.
+
+Setup
+-----
+
+An OTLP compatible endpoint should be listening for traces and logs on
+http://localhost:4317. If not, update "OTEL_EXPORTER_OTLP_ENDPOINT" as well.
+
+Next, set up a virtual environment like this:
+
+::
+
+    python3 -m venv .venv
+    source .venv/bin/activate
+    pip install "python-dotenv[cli]"
+    pip install -r requirements.txt
+
+You will also need an Anthropic API key. Set it as an environment variable:
+
+::
+
+    export ANTHROPIC_API_KEY=your_api_key_here
+
+Run
+---
+
+Run the example with zero-code instrumentation like this:
+
+::
+
+    dotenv run -- opentelemetry-instrument python main.py
+
+You should see a poem generated by Claude while traces and logs export to your
+configured observability tool. No changes to `main.py` were required!
+
+Learn More
+----------
+
+See the `OpenTelemetry Python automatic instrumentation docs
+<https://opentelemetry.io/docs/languages/python/automatic/>`_ for more
+information about zero-code instrumentation.
+
