+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/requirements.oldest.txt
@@ -1,0 +1,29 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This variant of the requirements aims to test the system using
+# the oldest supported version of external dependencies.
+
+-e util/opentelemetry-util-genai
+anthropic==0.16.0
+httpx>=0.25.2,<0.28.0  # Pin to version compatible with anthropic 0.16.0 (proxies arg removed in 0.28)
+pytest==7.4.4
+pytest-vcr==1.0.2
+pytest-asyncio==0.21.0
+wrapt==1.16.0
+opentelemetry-api==1.37  # when updating, also update in pyproject.toml
+opentelemetry-sdk==1.37  # when updating, also update in pyproject.toml
+opentelemetry-semantic-conventions==0.58b0  # when updating, also update in pyproject.toml
+
+-e instrumentation-loongsuite/loongsuite-instrumentation-anthropic
