+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/test_instrumentor.py
@@ -1,0 +1,128 @@
+# Copyright The OpenTelemetry Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for the AnthropicInstrumentor class."""
+
+from opentelemetry.instrumentation.anthropic import AnthropicInstrumentor
+
+
+def test_instrumentor_instantiation():
+    """Test that the instrumentor can be instantiated."""
+    instrumentor = AnthropicInstrumentor()
+    assert instrumentor is not None
+    assert isinstance(instrumentor, AnthropicInstrumentor)
+
+
+def test_instrumentation_dependencies():
+    """Test that instrumentation dependencies are correctly reported."""
+    instrumentor = AnthropicInstrumentor()
+    dependencies = instrumentor.instrumentation_dependencies()
+
+    assert dependencies is not None
+    assert len(dependencies) > 0
+    assert "anthropic >= 0.16.0" in dependencies
+
+
+def test_instrument_uninstrument_cycle(
+    tracer_provider, logger_provider, meter_provider
+):
+    """Test that instrument() and uninstrument() can be called multiple times."""
+    instrumentor = AnthropicInstrumentor()
+
+    # First instrumentation
+    instrumentor.instrument(
+        tracer_provider=tracer_provider,
+        logger_provider=logger_provider,
+        meter_provider=meter_provider,
+    )
+
+    # First uninstrumentation
+    instrumentor.uninstrument()
+
+    # Second instrumentation (should work)
+    instrumentor.instrument(
+        tracer_provider=tracer_provider,
+        logger_provider=logger_provider,
+        meter_provider=meter_provider,
+    )
+
+    # Second uninstrumentation
+    instrumentor.uninstrument()
+
+
+def test_multiple_instrumentation_calls(
+    tracer_provider, logger_provider, meter_provider
+):
+    """Test that multiple instrument() calls don't cause issues."""
+    instrumentor = AnthropicInstrumentor()
+
+    # First call
+    instrumentor.instrument(
+        tracer_provider=tracer_provider,
+        logger_provider=logger_provider,
+        meter_provider=meter_provider,
+    )
+
+    # Second call (should be idempotent or handle gracefully)
+    instrumentor.instrument(
+        tracer_provider=tracer_provider,
+        logger_provider=logger_provider,
+        meter_provider=meter_provider,
+    )
+
+    # Clean up
+    instrumentor.uninstrument()
+
+
+def test_uninstrument_without_instrument():
+    """Test that uninstrument() can be called without prior instrument()."""
+    instrumentor = AnthropicInstrumentor()
+
+    # This should not raise an error
+    instrumentor.uninstrument()
+
+
+def test_instrument_with_no_providers(
+    tracer_provider, logger_provider, meter_provider
+):
+    """Test that instrument() works without explicit providers.
+
+    Note: We still pass providers to ensure a clean test environment,
+    but this tests that the instrumentor can be called and cleaned up.
+    In a real scenario without explicit providers, it would use the
+    global (no-op) providers.
+    """
+    instrumentor = AnthropicInstrumentor()
+
+    # Test that instrument/uninstrument cycle works
+    instrumentor.instrument(
+        tracer_provider=tracer_provider,
+        logger_provider=logger_provider,
+        meter_provider=meter_provider,
+    )
+
+    # Clean up
+    instrumentor.uninstrument()
+
+
+def test_instrumentor_has_required_attributes():
+    """Test that the instrumentor has the required methods."""
+    instrumentor = AnthropicInstrumentor()
+
+    assert hasattr(instrumentor, "instrument")
+    assert hasattr(instrumentor, "uninstrument")
+    assert hasattr(instrumentor, "instrumentation_dependencies")
+    assert callable(instrumentor.instrument)
+    assert callable(instrumentor.uninstrument)
+    assert callable(instrumentor.instrumentation_dependencies)
