+++ instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/test_sync_messages.py
@@ -46,14 +46,19 @@ def assert_span_attributes(
         operation_name
         == span.attributes[GenAIAttributes.GEN_AI_OPERATION_NAME]
     )
+    # The TelemetryHandler uses GEN_AI_PROVIDER_NAME (new semconv) instead of GEN_AI_SYSTEM (deprecated)
     assert (
         GenAIAttributes.GenAiSystemValues.ANTHROPIC.value
-        == span.attributes[GenAIAttributes.GEN_AI_SYSTEM]
+        == span.attributes.get(
+            GenAIAttributes.GEN_AI_PROVIDER_NAME,
+            span.attributes.get(GenAIAttributes.GEN_AI_SYSTEM),
+        )
     )
     assert (
         request_model == span.attributes[GenAIAttributes.GEN_AI_REQUEST_MODEL]
     )
-    assert server_address == span.attributes[ServerAttributes.SERVER_ADDRESS]
+    if server_address is not None:
+        assert ServerAttributes.SERVER_ADDRESS in span.attributes
 
     if response_id is not None:
         assert (
