+++ tox-loongsuite.ini
@@ -24,6 +24,10 @@ envlist =
     py3{10,11,12,13}-test-loongsuite-instrumentation-claude-agent-sdk-{oldest,latest}
     lint-loongsuite-instrumentation-claude-agent-sdk
 
+    ; loongsuite-instrumentation-anthropic
+    py3{9,10,11,12,13}-test-loongsuite-instrumentation-anthropic-{oldest,latest}
+    lint-loongsuite-instrumentation-anthropic
+
     ; loongsuite-instrumentation-google-adk
     py3{9,10,11,12,13}-test-loongsuite-instrumentation-google-adk-{oldest,latest}
     lint-loongsuite-instrumentation-google-adk
@@ -105,6 +109,11 @@ deps =
   claude-agent-sdk-latest: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-claude-agent-sdk/tests/requirements.latest.txt
   lint-loongsuite-instrumentation-claude-agent-sdk: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-claude-agent-sdk/tests/requirements.oldest.txt
 
+  anthropic-oldest: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/requirements.oldest.txt
+  anthropic-latest: {[testenv]test_deps}
+  anthropic-latest: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/requirements.latest.txt
+  lint-loongsuite-instrumentation-anthropic: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests/requirements.oldest.txt
+
   google-adk-oldest: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-google-adk/tests/requirements.oldest.txt
   google-adk-latest: {[testenv]test_deps}
   google-adk-latest: -r {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-google-adk/tests/requirements.latest.txt
@@ -179,6 +188,9 @@ commands =
   
   test-loongsuite-instrumentation-claude-agent-sdk: pytest {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-claude-agent-sdk/tests {posargs}
   lint-loongsuite-instrumentation-claude-agent-sdk: python -m ruff check {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-claude-agent-sdk
+
+  test-loongsuite-instrumentation-anthropic: pytest {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-anthropic/tests {posargs}
+  lint-loongsuite-instrumentation-anthropic: python -m ruff check {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-anthropic
 
   test-loongsuite-instrumentation-google-adk: pytest {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-google-adk/tests {posargs}
   lint-loongsuite-instrumentation-google-adk: python -m ruff check {toxinidir}/instrumentation-loongsuite/loongsuite-instrumentation-google-adk
